draw helper as base utility
git-svn-id: svn://127.0.0.1/repos/trunk@827 a5c22d9f-9f17-4845-9224-d5bed68bb274

+++ myd3dlib/include/myFont.h
@@ -136,7 +136,7 @@ namespace my
 
 		my::Vector2 m_Scale;
 
-		float m_LineHeight;
+		int m_LineHeight;
 
 		CharacterMap m_characterMap;
